+++ app/src/main/java/com/reno/philipshuesampleapp/MainActivity.kt
@@ -10,6 +10,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.reno.philipshue.bridge.IDiscoveryManager
 import com.reno.philipshue.bridge.NUPnPDiscoveryManager
+import com.reno.philipshue.bridge.SocketDiscoveryManager
 import com.reno.philipshue.bridge.UPnPDiscoveryManager
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -39,34 +40,9 @@ class MainActivity : AppCompatActivity() {
         mAdapter = MyAdapter(myDataSet)
         mRecyclerView.adapter = mAdapter
 
-//        val uPnPManager = UPnPDiscoveryManager.Builder(this).build()
-//        uPnPManager.discoverDevices(object: IDiscoveryManager.DiscoverListener {
-//            override fun onStart() {
-//                Log.d(TAG, "Start discovery")
-//            }
-//
-//            override fun onFoundNewDevice(device: UPnPDevice) {
-//                Log.d(TAG, "Found new device")
-//                Log.d("App", device.location)
-//                myDataSet.add(device.toString())
-//                (mAdapter as MyAdapter).notifyDataSetChanged()
-//            }
-//
-//            override fun onFinish(devices: HashSet<UPnPDevice>) {
-//                // To do something
-//                Log.d(TAG,"Discovery finished")
-//            }
-//
-//            override fun onError(exception: Exception) {
-//                Log.d(TAG, "Error: " + exception.localizedMessage)
-//                exception.printStackTrace()
-//            }
-//
-//        })
-
-        val uPnPDiscoveryManager = UPnPDiscoveryManager.Builder(this).build()
+        val uPnPDiscoveryManager = UPnPDiscoveryManager(SocketDiscoveryManager.Builder(this).build())
         CoroutineScope(Dispatchers.IO).launch {
-            val uPnPDevices = uPnPDiscoveryManager.discoverDevices().toList()
+            val uPnPDevices = uPnPDiscoveryManager.getBridges()
             uPnPDevices.forEach { myDataSet.add(it.toString()) }
         }
