+++ philipshue/build.gradle
@@ -50,6 +50,7 @@ dependencies {
     testImplementation "org.koin:koin-test:$koin_version"
     // Koin for Java developers
     implementation "org.koin:koin-java:$koin_version"
+    implementation "org.koin:koin-android:$koin_version"
 
     testImplementation "io.mockk:mockk:1.9.3"
     testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.2"
